Rename stripe_checkout -> stripe_subscriptions.
